Create data generator to split responsibility
