+++ Assets/Scripts/Gameplay/UI/Tooltips/ActionTooltipDataGenerator.cs
@@ -1,0 +1,82 @@
+using Unity.BossRoom.Gameplay.Actions;
+
+namespace Unity.BossRoom.Gameplay.UI
+{
+    public static class ActionTooltipDataGenerator
+    {
+        public static TooltipData Generate(ActionConfig config)
+        {
+            var rootText = string.Format(TooltipSettings.Default.TooltipFormat, config.DisplayedName,
+            string.Format(config.Description, config.Logic.ToString()));
+
+            var rootTooltip = new TooltipData(rootText);
+            var tooltipChain = CreateTooltipChain(config);
+
+            rootTooltip.Attach(tooltipChain);
+
+            return rootTooltip;
+        }
+
+        private static string InjectConfigIntoTemplate(ActionConfig config, int projectileIndex = 0)
+        {
+            if (!TooltipSettings.Default.TooltipDatabase.TryGetTooltipLinkData(config.Logic.ToString(), out var template))
+            {
+                return config.DisplayedName;
+            }
+
+            return ActionTooltipInjector
+                .InjectIntoTemplate(template.Description, config, projectileIndex);
+        }
+
+        private static TooltipData CreateTooltipChain(ActionConfig config)
+        {
+            TooltipData rootTooltip = null;
+
+            if (config.Projectiles.Length > 1)
+            {
+                var currentTooltip = rootTooltip;
+
+                for (int i = 0; i < config.Projectiles.Length; i++)
+                {
+                    var nextTooltip = GetDataFromProjectile(config, i);
+                    if (rootTooltip == null)
+                    {
+                        rootTooltip = nextTooltip;
+                        currentTooltip = nextTooltip;
+                    }
+                    else
+                    {
+                        currentTooltip.Attach(nextTooltip);
+                        currentTooltip = nextTooltip;
+                    }
+                }
+            }
+            else
+            {
+                string configTooltipText = InjectConfigIntoTemplate(config);
+                rootTooltip = new TooltipData(configTooltipText);
+            }
+
+            return rootTooltip;
+        }
+
+        private static TooltipData GetDataFromProjectile(ActionConfig config, int i)
+        {
+            string projectileTooltipText = InjectConfigIntoTemplate(config, i);
+
+            if (i < config.Projectiles.Length - 1)
+            {
+                var nextProjectile = config.Projectiles[i + 1];
+                string linkText = GetNextProjectileLink(nextProjectile, config);
+                projectileTooltipText = $"{projectileTooltipText} {linkText}";
+            }
+
+            return new TooltipData(projectileTooltipText);
+        }
+
+        private static string GetNextProjectileLink(ProjectileInfo nextProjectile, ActionConfig config)
+        {
+            return $"Turns into <link={config.Logic}><style=Clickable>{nextProjectile.ProjectilePrefab.name}</link></style> projectile upon {config.DurationSeconds / config.Projectiles.Length} seconds.";
+        }
+    }
+}
