+++ Assets/Scripts/Gameplay/UI/Tooltips/ActionTooltipDataGenerator.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d269e63155912ed4d814be3dc01e7fc4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
